+++ android/settings.gradle
@@ -1,4 +1,4 @@
-rootProject.name = 'ReactNativeTemplate'
+rootProject.name = 'BloomMate'
 apply from: file("../node_modules/@react-native-community/cli-platform-android/native_modules.gradle"); applyNativeModulesSettingsGradle(settings)
 include ':app'
 includeBuild('../node_modules/@react-native/gradle-plugin')
