+++ app.json
@@ -1,4 +1,4 @@
 {
-  "name": "ReactNativeTemplate",
-  "displayName": "ReactNativeTemplate"
+  "name": "BloomMate",
+  "displayName": "BloomMate"
 }
