--- ios/ReactNativeTemplate/AppDelegate.h
@@ -1,6 +1,0 @@
-#import <RCTAppDelegate.h>
-#import <UIKit/UIKit.h>
-
-@interface AppDelegate : RCTAppDelegate
-
-@end
