+++ package.json
@@ -1,5 +1,5 @@
 {
-  "name": "ReactNativeTemplate",
+  "name": "bloommate",
   "version": "0.0.1",
   "private": true,
   "scripts": {
